refactor: 472 - rename events folder
