+++ src/entities/event/index.ts
