+++ src/entities/event/types.ts
