+++ src/entities/event/ui/event-card/event-card.module.scss
