+++ src/entities/event/ui/event-card/event-card.test.tsx
